requirements.txt
APScheduler==3.9.1.post1
asgiref==3.6.0
black==22.12.0
click==8.1.3
Django==4.1.5
feedparser==6.0.10
mypy-extensions==0.4.3
pathspec==0.10.3
platformdirs==2.6.2
python-dateutil==2.8.2
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
sgmllib3k==1.0.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
tzlocal==4.2
